Standardized on float for camera.
That's what all other classes in my code use too, after all.

+++ GLLara/GLLCamera.h
@@ -16,33 +16,33 @@
 @interface GLLCamera : NSManagedObject
 
 @property (nonatomic) BOOL cameraLocked;
-@property (nonatomic) double distance;
-@property (nonatomic) double farDistance;
-@property (nonatomic) double fieldOfViewY;
+@property (nonatomic) float distance;
+@property (nonatomic) float farDistance;
+@property (nonatomic) float fieldOfViewY;
 @property (nonatomic) int64_t index;
-@property (nonatomic) double latitude;
-@property (nonatomic) double longitude;
-@property (nonatomic) double nearDistance;
-@property (nonatomic) double positionX;
-@property (nonatomic) double positionY;
-@property (nonatomic) double positionZ;
-@property (nonatomic) double windowHeight;
+@property (nonatomic) float latitude;
+@property (nonatomic) float longitude;
+@property (nonatomic) float nearDistance;
+@property (nonatomic) float positionX;
+@property (nonatomic) float positionY;
+@property (nonatomic) float positionZ;
+@property (nonatomic) float windowHeight;
 @property (nonatomic) BOOL windowSizeLocked;
-@property (nonatomic) double windowWidth;
+@property (nonatomic) float windowWidth;
 @property (nonatomic, retain) GLLCameraTarget *target;
 
 // Transient
-@property (nonatomic) double actualWindowHeight;
-@property (nonatomic) double actualWindowWidth;
+@property (nonatomic) float actualWindowHeight;
+@property (nonatomic) float actualWindowWidth;
 
 // Completely independent
-@property (nonatomic) double latestWindowHeight;
-@property (nonatomic) double latestWindowWidth;
+@property (nonatomic) float latestWindowHeight;
+@property (nonatomic) float latestWindowWidth;
 
 // Derived
-@property (nonatomic) double currentPositionX;
-@property (nonatomic) double currentPositionY;
-@property (nonatomic) double currentPositionZ;
+@property (nonatomic) float currentPositionX;
+@property (nonatomic) float currentPositionY;
+@property (nonatomic) float currentPositionZ;
 
 @property (nonatomic, readonly) mat_float16 viewProjectionMatrix;
 @property (nonatomic, readonly) vec_float4 cameraWorldPosition;
